Transition to Abacus-themed env var prefix.
Previously we had lots of environment variables prefixed with OPT
as a holdover from the Optics fork. This PR tries to migrate them to
Abacus-specific prefix of ABC, which is something we'd like to do
prior to handing operational responsibility for running agents to
any third parties, since there will be some inherent resistance in
persuading a group like that to rename all configuration variables.