Provide inbox signer to GelatoSubmitter init.
This might fail, in which case the relayer can't really
do anything for that inbox if it is configured to use
Gelato. In that case, bail!(..). Bailing requires that the
boot/init boilerplate in the relayer handle errors, so
plumb/wrap those all in Result<> return types and update
accordingly.