Reduce log severity for already-sent messages. (#735)
It is expected for the 'processor' module to churn through the entire
history of sent messages on relayer init, and it is expected/normal
behavior to discover that the vast majority of those have already been
sent. Therefore, it is cluttering / inappropriate to log those events at
info! severity, and instead preferable to record them at debug severity.

Co-authored-by: Mattie Conover <git@mconover.dev>