run-locally: Run 'yarn prettier' after hardhat.
Currently the process of starting/running our local hardhat
setup results in an annoying irrelevant diff created
against a test.json config file.

A better version of this change would probably fix the
typescript script that touches / generates that file, but
I'm completely unfaimilar with that setup (and ts in
general) so opting for this hack for now.