run-locally: Run 'yarn prettier' after hardhat. (#750)
Currently the process of starting/running our local hardhat
setup results in an annoying irrelevant diff created
against a test.json config file. The diff is due to lack of
a file-terminating newline character.

A better version of this change would probably fix the
typescript script that touches / generates that file, but
I'm completely unfamiliar with that setup (and TS in
general) so opting for this hack for now.