Fix broken test `contract_sync::outbox::test::handles_missing_rpc_messages`.
Before this PR, the test will panic due to outbox syncer running in a concurrent
tokio task after the test completes and invoking mocked functions without expectations set.
The panic is also swallowed by a separate bug and is only visible if tests are run under
`cargo test -- --show-output`.

After this PR, we no longer see a backtrace / panic when running with `cargo test -- --show-output`.
We also correctly mark the test as failed in `cargo test` if we introduce a new panic in the test or DB harness code.
Some additional changes to try and improve things adjacent to this test:
    *  use tmpfs instead of disk + panic-unwind-detection for the RocksDB contents to avoid the panic detection complexity
    *  avoid the current correctness dependency on sleep(from_secs(5)) in the test. This is inherently racy and it is not much harder or more complex to just select() with a timeout