Fixes #786
Adds arbitrum, bsc, and optimism and their domains to the `domain_and_chain!` macro, which will enable portions of our monitoring to properly identify traffic relating to this chains instead of filling with "unknown" today.

Also introduces a unit test that searches for any json config files it can find, tries to parse them all, takes the union of all chain names and domain mappings it can see, and ensures certain internal consistency properties hold and that the `domain_and_chain!` macro is aware of them.

It's possible that this could be eventually adapated for json configs living in `typescript/infra` instead of `rust` if we go that direction. My hope was just to avoid a situation where we forget to keep things synchronized again.