Remove gelato support (#1946)
### Description

We have not been using Gelato anymore so thought it would be nice to
remove this code and avoid continuing to maintain it. I know I have had
to update it a few times to maintain parity. As an added perk it should
slightly speed up build times since it removes an entire crate.

### Drive-by changes

None

### Related issues


### Backward compatibility

_Are these changes backward compatible?_

Yes - As we do not use this code

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

Unit Tests