Better Agent Configuration Parsing (#2070)
### Description

This is a significant change to how we parse user configs. The main goal
of this PR is to give a clear path and env key to any errors that are
encountered when parsing configuration.

In order to accomplish this, I have crated a new trait `FromRawConf`
which defines a set of steps to validate and convert individual fields.
Also for every major config struct, there is now a `Raw*` variant of it
which is used by the `config` library to reduce errors that occur during
deserialization itself. `Raw*` types should always use the most basic
form of a value, such as `String`, `StrOrInt`, `bool`, and ideally
optional in all cases. Missing values should be handled during the raw
conversion and not by `config`.

I also needed to make changes to a number of types stored in the parsed
config types to force validation forward to the parsing step instead of
doing it when we read the config value. This also required me to create
a filter to prevent trying to validate certain configs that we would not
ever need to load.

These changes can also be built on later to support something other than
`config` if we choose to, or add support for merging configs from
multiple sources since everything is optional.

### Drive-by changes

- Default to `http` for connection type
- Default to no gas payment enforcement if not specified instead of
failing
- `ChainSetup` -> `ChainConf`
- `GasPaymentEnforcementConfig` -> `GasPaymentEnforcementConf`
- Made ethereum connection configuration more forgiving
- Moved hyperlane base settings from `mod.rs` to `base.rs`
- Moved config chain tests in hyperlane core to `tests` dir to fix a
cyclical import problem
- Extension traits to help with config in hyperlane core
- Moved `StrOrInt` to new hyperlane core `config` module
- Support for parsing `U256` from `StrOrInt`
- Removed `HexString` type which is now redundant
- Updated base settings to use hyperlane domain
- Use `heyKey` as signer type if `type` is not specified and `key` is
- Moved hyperlane ethereum chain config to a new module

### Related issues

- Fixes #2033 
- Fixes #2012 

### Backward compatibility

_Are these changes backward compatible?_

Yes - This should take in configs in the same shape but be more
forgiving in a few places

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

Manual
Unit Test