New deploy scripts, remove canonical IDs for our programs (#2389)
### Description

* Adds `core deploy` and `warp-route deploy` scripts into the client.
There's still a lot to be desired here to be totally honest and it's not
super clean/robust - this is a symptom of trying to get some results
* Moves bash script over to using these new scripts
* Deployments are made to an environment and chain. Warp route
deployments similarly have an identifying name. See
`rust/sealevel/environments` for these artifacts. Keys for programs are
created or reused if present already and reuse is requested. See the
bash script for examples
* Includes a deploy I made to the Solana devnet as well as artifacts /
keys for the local-e2e environment
* Moves away from canonical IDs for all our core and warp route programs
* Removes `recipient/echo` and the rubber stamp ISM
* Moves to a new version of SPL libs that use their canonical IDs (these
were changed in the fork so that they could be deployed with a key we
have access to, but we can instead just run `solana-test-validator` with
programs from genesis

### Drive-by changes

Cleanup / test refactors as required

### Related issues

- Fixes #2349 
- Fixes #2288 (sorta :P)
- Partially #2289 - I did deploy to the Solana devnet, not Nautilus.
Will likely wanna do another deploy in the future?
- Partially #2366 - need to make some changes to the EVM tooling side to
be able to enroll Sealevel routers

### Backward compatibility

_Are these changes backward compatible?_

New local e2e instructions

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

Unit Tests, local e2e