Reconstruct relayer `OpQueue` from HyperlaneDB (#2382)
### Description

Stores the `num_retries` field of `PendingMessage` at its own prefix in
HyperlaneDB, so that `PendingMessage` can be reconstructed with the same
exponential backoff in the `next_attempt_after` field and avoid making
unnecesary RPC calls on restart.

To use persisted data, `PendingMessage` is constructed with
`from_persisted_retries` instead of the default constructor. If reading
from the db fails, the new constructor just defaults to the regular
constructor, so it's worth considering replacing the regular constructor
with this one entirely.

### Drive-by changes

A new db prefix is added (`PENDING_MESSAGE_RETRY_COUNT_FOR_MESSAGE_ID
`), as well as related getter / setter methods. Using the new prefix
helps avoid a migration. I do see a `processed_by_nonce` prefix, which
could be migrated to store data related to `PendingMessage` in general
so let me know what you think. Could be a future improvement to merge
its prefix with the prefix added in this PR.

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/1902

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

A processor is run on some dummy database entries. The
`SerialSubmitter`'s end of the channel is used to ensure that the
incoming `PendingMessage`s have the correct backoff in
`_next_attempt_after`.