feat(fallback provider): re-prioritize providers on error (#2369)
### Description

Closes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/2189

Reprioritize providers in the `FallbackProvider` when they stop
receiving new blocks (although a `max_block_time` has elapsed). The
`JsonRpcClient::request(...)` implementation now takes a snapshot of the
provider priority list and iterates on that. This is to minimize lock
contention - no lock is held across an `await` so there is no deadlock
risk. If one of the errors above occurs, the priority list is
overwritten.

Note this may cause a prior reprioritization to be lost, if the
repriotization occured after the current tokio task's snapshot was
taken. I'm not entirely sure whether a specific `FallbackProvider` is
referenced by more than one task at a time though.

### Drive-by changes

A `PrometheusJsonRpcClientConfigExt` trait is added, so the provider can
be mocked.

### Testing

Unit tests.