Final cleanup, make process nonreentrant (#2459)
### Description

Took a pass everywhere where there were "FIXME"s or "TODO"s. This is
mostly a cosmetic change -- most changes are just adding `///` docs, or
shuffling things around.

Things that actually changed are:

1. Moved the `SizedData` trait and `AccountData` to `account_utils`,
instead of the Mailbox program
2. Made the `InboxProcess` instruction on the Mailbox effectively
non-reentrant. This is done by holding a `RefMut` of the Inbox PDA
account's data. Added a test for this too.
3. Removed the half-done "event" logging on hyperlane-token. We may want
to consider something like this in a future iteration where we will
expect to do some more serious indexing of events, but for now we don't
need this anywhere.

### Drive-by changes

docs

### Related issues

- Fixes #2353
- Fixes #2218

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None

### Testing

_What kind of testing have these changes undergone?_

Unit Tests, e2e test