Support 64-byte txn ids (attempt 2) (#2612)
### Description

Adds support to TxOutcome for 512-bit ids and to log meta to support
Solana. Now that we have primitive types in our crate we were able to
create a better solution than before which just stores the EVM 256bit
hashes inside of the 512bit one with leading zeros rather than needing
to over engineer an enum container for it.

### Drive-by changes

- Removed some dead code
- Improved primitive type support for 512bit types
- Improved serialization support for 512bit types

### Related issues

- Fixes #2252 
- Replaces #2256

### Backward compatibility

_Are these changes backward compatible?_

Yes-ish, changes one of the internal database key strings. At standup we
discussed the possibility of migrating the data and concluded that was a
can of worms we did not want to open.

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None

### Testing

_What kind of testing have these changes undergone?_

Unit tests