Agent Config Base Parsing (#2665)
### Description

This is a big one for #2215...

1. Splits the parsing of configs from the config definitions to support
both the old, new, and future parsers simultaneously. (We might want to
rename it from parser to config format or something; not set on this
name).
2. Implements the new config format without switching to it or enabling
it. This means this PR does introduce dead code that is going to be used
in the next PR.

### Drive-by changes

- Deleted some unused (or single use) macros defined in hyperlane base
- Updated the definition of the config parsing trait to simplify
implementation
- Added some more ergonomic methods of handling errors in config parsing
- Removed any raw `unwrap()` I noticed since these cause difficult to
debug errors in prod.
- Added some additional validation to catch code mistakes in config path
definitions

### Related issues

- Progress on #2215
- Comes after #2658

### Backward compatibility

Yes

### Testing
Manual (E2E only for now to make sure I did not break the old config
format)