feat: Add possibility to configure a subfolder inside an S3 bucket (#2647)
### Description

This change adds the possibility to configure a subfolder inside an S3
bucket, to avoid having to spin up an entire bucket for hyperlane.

### Drive-by changes

Nothing

### Related issues

None, chatted about it on Slack with @nambrot

### Backward compatibility

Yes

I made the field optional in the `RawCheckpointSyncerConf` so that
current config are still valid and will continue to read/write at the
root of the bucket. That being said, I'm new to this code so I hope I
got it right

### Testing

None

---------

Co-authored-by: Nam Chu Hoai <nambrot@googlemail.com>

(cherry picked from commit 478826b3a0adbb3637ba0051d8c77c0990d063f6)