Sealevel IGP debugging, temporarily change Sealevel gas payment policy (#2758)
### Description

- Some small changes to help with debugging sealevel IGP
- Turns out the GasPaymentAccount data format onchain is different from
the one in `main`. This causes IGP indexing to not work in the wild, but
it works in e2e. Because we were planning to change the IGP anyways as
we transfer ownership to Zebec, this isn't so bad - so temporarily just
not enforcing ZBC warp route Solana -> Nautilus gas payments. In
practice, we're still effectively enforcing gas payments because the
warp route will pay for gas

### Drive-by changes

- Also added the new BSC POSE addy

### Related issues

n/a

### Backward compatibility

yes

### Testing

Deployed, poked around, used cmds