Streamline Rust Agent Config Shapes (#2659)
Continues the updates to the rust config shapes by updating deployment
and runtime expectations. This PR also attempts to rely on the new
single source of truth created by the schema in the SDK as much as
reasonably possible which helped delete more code but also give some
guarantees of consistency.

THIS IS A BREAKING CHANGE! It changes the config shapes the agents want
and we should not merge this until we are ready.

Fixes #2215

---------

Co-authored-by: Guillaume Bouvignies <guillaumebouvignies@gmail.com>
Co-authored-by: Yorke Rhodes <yorke@hyperlane.xyz>
Co-authored-by: Guillaume Bouvignies <guillaume.bouvignies@kurtosistech.com>