refactor: finish removing unwraps (#2881)
### Description

- Removes unwraps and expects from everywhere but `run-locally`
- Precomputes the `address` in `Signer` at construction time, to
propagate the error early and keep signatures ergonomic by not requiring
a `Result`. Couldn't also precompile `SigningKey` (the privkey type)
because it's not `Sync` 😢
- Defines a `hyperlane-cosmos`-specific error enum
(`HyperlaneCosmosError`), which can be converted to
`ChainCommunicationError` with the `?` operator. This is a pattern we'd
like to refactor towards in the future, to remove dependencies from
`hyperlane-core` as much as possible
- One inconvenience is that you need to `.map_err()` to
`HyperlaneCosmosError` first, to use `?`. I wish `?` had deref coercion
semantics where it'd keep covnerting until an error matches, but I
assume this isn't possible because while you can only have a single
`Deref` impl, you can have multiple `From<Err>` impls.
- Writing this I'm realising we could write a small macro to implement
`From<Err> for ChainCommunicationError` for all the sub-errors of
`HyperlaneCosmosError` et al (cc @tkporter)