Relayer balance metrics (#2976)
Done:
- scaffolding for fetching custom agent metrics
- abstractions for building a metrics fetcher for a given VM
- querying cosmos balances; e2e tested.
- querying evm balances; e2e tested.
- **Note that as a result, evm addresses are now no longer zero-padded
when printed in the logs. This may break existing log queries**
- fixed a nasty bug on ubuntu where wasmd (osmosisd dependency, part of
the grpc query flow) would panic when a block is specified via
`x-cosmos-block-height`. The fix was to bump the version of osmosisd
from `19.0.0` to `20.5.0`. **Note that when running e2e on Mac OS, the
osmosis version in use is still 19.0.0**. That's because we need a fork
that publishes a darwin target binary (currently pointing
[here](https://github.com/hashableric/osmosis/releases/download/v19.0.0-mnts/osmosisd-19.0.0-mnts-darwin-arm64.tar.gz))

For follow up PR:
- sealevel balance querying

I'm open to all renaming suggestions, I just tried to speed through and
didn't ponder names too much

Relates to https://github.com/hyperlane-xyz/issues/issues/701
Closes https://github.com/hyperlane-xyz/issues/issues/702 (because the
balance becomes available in the metrics endpoint for polling)